添加获取云盘音乐接口

+++ src/main/kotlin/com/dai/bean/music/SongList.kt
@@ -3,7 +3,7 @@ package com.dai.bean.music
 
 //所有人员歌单列表
 data class SongList(
-        val personId: Long, //人员id
+        val personId: Int, //人员id
         val name: String, //歌单名字
         val createTime: String, //创建时间
         val introduction: String  //歌单简介
